requirements.txt
flask
sqlalchemy
flask_sqlalchemy
jinja2
python-dotenv
requests
Werkzeug
Flask-Limiter